moved worklist to own header, put some stuff in nicer macros
