+++ src/runtime/memory/allocator.c
@@ -10,7 +10,7 @@ size_t root_count = 0;
 AllocatorBin a_bin = {.freelist = NULL, .entrysize = DEFAULT_ENTRY_SIZE, .page = NULL, .page_manager = NULL};
 PageManager p_mgr = {.all_pages = NULL, .evacuate_page = NULL, .filled_pages = NULL};
 
-/* To be used with updating pointers after moving to evac page */
+/* Forwarding table to be used with updating pointers after moving to evac page */
 Worklist f_table = {.size = 0};
 
 /* Our stack of roots to be marked after allocations finish */
@@ -123,88 +123,29 @@ bool isRoot(void* obj)
     return true; // For now, assume all objects are valid pointers
 }
 
-/* Worklist helpers */
-static void initialize_worklist(Worklist* worklist) 
-{
-    worklist->size = 0;
-}
-
-#if 0
-static Object* worklist_top(Worklist* worklist)
-{
-    assert(worklist->size > 0);
-
-    if(worklist->size > 0) {
-        return worklist->data[worklist->size - 1];
-    } else {
-        return worklist->data[0];
-    }
-}
-#endif
-
-static bool add_to_worklist(Worklist* worklist, Object* obj) 
-{
-    if (worklist->size >= WORKLIST_CAPACITY) {
-        /* Worklist is full */
-        debug_print("Worklist overflow!\n");
-        return false;
-    }
-    worklist->data[worklist->size++] = obj;
-    return true;
-}
-
-static Object* remove_from_worklist(Worklist* worklist) 
-{
-    if (worklist->size == 0) {
-        return NULL;
-    }
-    return worklist->data[--worklist->size]; //prefix decrement crucial here
-}
-
-// I really feel that there are better choices than this
-static Object* remove_oldest_from_worklist(Worklist* worklist) {
-    if (worklist->size == 0) {
-        return NULL;
-    }
-    
-    Object* oldest = worklist->data[0];
-    
-    // Shift all elements to the left
-    for (size_t i = 1; i < worklist->size; i++) {
-        worklist->data[i - 1] = worklist->data[i];
-    }
-    
-    worklist->size--;
-    return oldest;
-}
-
-static bool is_worklist_empty(Worklist* worklist) 
-{
-    return worklist->size == 0;
-}
-
 // Move object to evacuate_page and reset old metadata
 static void* evacuate_object(AllocatorBin *bin, Object* obj, Worklist* forward_table) {
     FreeListEntry* start_of_evac = bin->page_manager->evacuate_page->freelist;
     FreeListEntry* next_evac_freelist_object = start_of_evac->next;
 
     // Direct copy of data from original page to evacuation page
-    void* evac_obj_base = memcpy(start_of_evac, (char*)obj - sizeof(MetaData) - ALLOC_DEBUG_CANARY_SIZE, REAL_ENTRY_SIZE(DEFAULT_ENTRY_SIZE));
+    void* evac_obj_base = memcpy(start_of_evac, BLOCK_START_FROM_OBJ(obj), REAL_ENTRY_SIZE(DEFAULT_ENTRY_SIZE));
 
     //update freelist of evac page, memcpy was destroying pointers in our freelist for evac page so I had to manually store and assign after memcpy
     bin->page_manager->evacuate_page->freelist = next_evac_freelist_object;
 
     debug_print("[DEBUG] Object moved from %p to %p in evac page\n", obj, evac_obj_base);
 
-    Object* evac_obj_data = (Object*)((char*)evac_obj_base + sizeof(MetaData) + ALLOC_DEBUG_CANARY_SIZE);
-    MetaData* evac_obj_meta = (MetaData*)((char*)evac_obj_data - sizeof(MetaData));
+    Object* evac_obj_data = (Object*)(OBJ_START_FROM_BLOCK(evac_obj_base));
+    MetaData* evac_obj_meta = META_FROM_OBJECT(evac_obj_data);
     
     // Insert evacuated object into forward table and set index in meta
     evac_obj_meta->forward_index = forward_table->size;
     add_to_worklist(forward_table, evac_obj_data);
 
     // Update the freelist of the original page to point to the newly freed block
-    FreeListEntry* new_freelist_entry = (FreeListEntry*)((char*)obj - sizeof(MetaData) - ALLOC_DEBUG_CANARY_SIZE);
+    // I suspect I could handle this freelist stuff within my rebuild freelist method.
+    FreeListEntry* new_freelist_entry = (FreeListEntry*)(BLOCK_START_FROM_OBJ(obj));
     new_freelist_entry->next = bin->freelist;
     bin->freelist = new_freelist_entry;
 
@@ -215,6 +156,14 @@ static void* evacuate_object(AllocatorBin *bin, Object* obj, Worklist* forward_t
     RESET_METADATA_FOR_OBJECT(old_metadata);
 
     return evac_obj_data;
+}
+
+/* Helper to update an objects children pointers */
+void update_children_pointers(Object* obj, Worklist* worklist, Worklist* forward_table) {
+    for (int i = 0; i < obj->num_children; i++) {
+        Object* oldest = remove_oldest_from_worklist(worklist);
+        obj->children[i] = forward_table->data[META_FROM_OBJECT(oldest)->forward_index];
+    }
 }
 
 void evacuate(Worklist *marked_nodes_list, AllocatorBin *bin) {
@@ -234,22 +183,13 @@ void evacuate(Worklist *marked_nodes_list, AllocatorBin *bin) {
                 add_to_worklist(&worklist, (Object*)evacuate_object(bin, obj, forward_table));
             } else {
                 // If the object we are trying to evacuate has children we will first update its children pointers then evacuate.
-                for (int i = 0; i < obj->num_children; i++) {
-                    // We want the OLDEST element in the worklist here -- not sure how to do so. It would also have to be removed.
-                    Object* oldest = remove_oldest_from_worklist(&worklist);
-                    obj->children[i] = forward_table->data[META_FROM_OBJECT(oldest)->forward_index];
-                }
+                update_children_pointers(obj, &worklist, forward_table);
                 add_to_worklist(&worklist, (Object*)evacuate_object(bin, obj, forward_table));
             }
         } else{
-            Object* root = remove_oldest_from_worklist(&worklist);
-            MetaData* root_meta = META_FROM_OBJECT(root);
-
             // If the object we are trying to evacuate has children we will first update its children pointers then evacuate.
-            for (int i = 0; i < obj->num_children; i++) {
-                obj->children[i] = forward_table->data[root_meta->forward_index];
-            }
-            add_to_worklist(&worklist, obj);
+            update_children_pointers(obj, &worklist, forward_table);
+            add_to_worklist(&worklist, obj); //Difference is we dont evac a root
         }
     }
 }
@@ -259,7 +199,7 @@ void rebuild_freelist(PageInfo* page) {
     page->freecount = 0;
 
     for(size_t i = 0; i < page->entrycount; i++) {
-        FreeListEntry* new_freelist_entry = (FreeListEntry*)((char*)page + sizeof(PageInfo) + (i * REAL_ENTRY_SIZE(page->entrysize)));
+        FreeListEntry* new_freelist_entry = (FreeListEntry*)(PAGE_OFFSET(page) + (i * REAL_ENTRY_SIZE(page->entrysize)));
         MetaData* meta = (MetaData*)((char*)new_freelist_entry + ALLOC_DEBUG_CANARY_SIZE);
 
         if(meta->isalloc == false) {
@@ -271,13 +211,13 @@ void rebuild_freelist(PageInfo* page) {
     }
 }
 
-// Just run through all pages and flag nodes that arent marked as non allocated
+// Just run through all pages and flag nodes that arent marked as non allocated --- may not be necessary
 void clean_nonref_nodes() {
     PageInfo* current_page = p_mgr.all_pages;
 
     while(current_page != NULL) {
         for(uint16_t i = 0; i < current_page->entrycount; i++) {
-            Object* current_object = (Object*)((char*)current_page + sizeof(PageInfo) + ALLOC_DEBUG_CANARY_SIZE + sizeof(MetaData) +
+            Object* current_object = (Object*)(PAGE_OFFSET(current_page) + ALLOC_DEBUG_CANARY_SIZE + sizeof(MetaData) +
                 (i * REAL_ENTRY_SIZE(current_page->entrysize)));
             MetaData* current_object_meta = META_FROM_OBJECT(current_object);
 
@@ -350,76 +290,4 @@ void mark_from_roots()
     evacuate(&marked_nodes_list, &a_bin);
 
     clean_nonref_nodes();
-}
-
-bool verifyCanariesInBlock(char* block, uint16_t entry_size)
-{
-    uint64_t* pre_canary = (uint64_t*)(block);
-    uint64_t* post_canary = (uint64_t*)(block + ALLOC_DEBUG_CANARY_SIZE + sizeof(MetaData) + entry_size);
-
-    if (*pre_canary != ALLOC_DEBUG_CANARY_VALUE || *post_canary != ALLOC_DEBUG_CANARY_VALUE)
-    {
-        debug_print("[ERROR] Canary corrupted at block %p\n", (void*)block);
-        debug_print("Data in pre-canary: %lx, data in post-canary: %lx\n", *pre_canary, *post_canary);
-        return false;
-    }
-    return true;
-}
-
-void verifyCanariesInPage(PageInfo* page)
-{
-    FreeListEntry* list = page->freelist;
-    char* base_address = (char*)page + sizeof(PageInfo);
-    uint16_t alloced_blocks = 0;
-    uint16_t free_blocks = 0;
-
-    for (uint16_t i = 0; i < page->entrycount; i++) {
-        char* block_address = base_address + (i * REAL_ENTRY_SIZE(page->entrysize));
-        debug_print("Checking block: %p\n", block_address);
-        MetaData* metadata = (MetaData*)(block_address + ALLOC_DEBUG_CANARY_SIZE);
-        debug_print("Metadata state: isalloc=%d\n", metadata->isalloc);
-
-        if (metadata->isalloc) {
-            alloced_blocks++;
-            assert(verifyCanariesInBlock(block_address, page->entrysize));
-        }
-    }
-
-    debug_print("\n");
-
-    while(list){
-        debug_print("Checking freelist block: %p\n", (void*)list);
-        MetaData* metadata = (MetaData*)((char*)list + ALLOC_DEBUG_CANARY_SIZE);
-        debug_print("Metadata state: isalloc=%d\n", metadata->isalloc);
-        if(metadata->isalloc){
-            debug_print("[ERROR] Block in free list was allocated\n");
-            assert(0);
-        }
-        free_blocks++;
-        list = list->next;
-    }   
-
-    // Make sure no blocks are lost
-    assert((free_blocks + alloced_blocks) == page->entrycount);
-
-    debug_print("\n");
-}
-
-void verifyAllCanaries(AllocatorBin* bin)
-{
-    PageInfo* current_page = bin->page_manager->all_pages;
-    PageInfo* evac_page = bin->page_manager->evacuate_page;
-
-    while (current_page) {
-        debug_print("PageManager all_pages head address: %p\n", current_page);
-        verifyCanariesInPage(current_page);
-        current_page = current_page->next;
-    }
-
-    while (evac_page) {
-        debug_print("PageManager evac_page head address: %p\n", evac_page);
-        verifyCanariesInPage(evac_page);
-        evac_page = evac_page->next;
-    }
-
 }
