+++ test/test.h
@@ -11,10 +11,15 @@ void test_mark_cyclic_graph(AllocatorBin* bin, PageManager* pm);
 void test_canary_failure(AllocatorBin* bin, PageManager* pm);
 void test_evacuation(AllocatorBin* bin, PageManager* pm);
 
-/**  
-* Lets add some more tests for lots of roots spanning multiple pages with (max?) children,
-* the introduction of cycles, and lots of objects not conntected to a root!
-**/
+/**
+ * Traverse pages and freelists ensuring no canaries are clobbered and that
+ * our freelists contain no already allocated objects.
+ **/
+#ifdef ALLOC_DEBUG_CANARY
+void verifyAllCanaries(AllocatorBin* bin);
+void verifyCanariesInPage(PageInfo* page);
+bool verifyCanariesInBlock(char* block, uint16_t entry_size);
+#endif
 
 /**
  * Run all tests together
